+++ components/src/core/drawer/drawer.tsx
@@ -1,20 +1,20 @@
 import React, { useCallback } from 'react';
-import { StyleSheet, ViewStyle, StyleProp, ViewProps } from 'react-native';
+import { StyleSheet, ViewStyle, ViewProps } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import { DrawerInheritableProps, inheritDrawerProps } from './inheritable-types';
-import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import { Surface, useTheme } from 'react-native-paper';
-import color from 'color';
 import { EdgeInsets } from '../__types__';
-import { $DeepPartial } from '@callstack/react-theme-provider';
+import { AllSharedProps } from './types';
+import { findChildByType, inheritSharedProps } from './utilities';
+import { DrawerContext } from './context';
 
+type DrawerStyles = {
+    root?: ViewStyle;
+};
 const makeStyles = (
     props: DrawerProps,
     theme: ReactNativePaper.Theme,
     insets: EdgeInsets
-): StyleSheet.NamedStyles<{
-    root: ViewStyle;
-}> =>
+): StyleSheet.NamedStyles<DrawerStyles> =>
     StyleSheet.create({
         root: {
             backgroundColor: props.backgroundColor || theme.colors.surface,
@@ -25,61 +25,58 @@ const makeStyles = (
         },
     });
 export type DrawerProps = ViewProps &
-    DrawerInheritableProps & {
-        style?: StyleProp<ViewStyle>;
+    AllSharedProps & {
+        // the id for the currently active item
+        activeItem?: string;
 
-        /**
-         * Overrides for theme
-         */
-        theme?: $DeepPartial<ReactNativePaper.Theme>;
+        // Function called whenever a navigation item or rail item is clicked
+        onItemSelect?: (id: string) => void;
+
+        // Custom style overrides
+        styles?: DrawerStyles;
     };
 export const Drawer: React.FC<DrawerProps> = (props) => {
-    const { theme: themeOverride, style, ...viewProps } = props;
-    const theme = useTheme(themeOverride);
-    // Nested expand/collapse icon defaults are different and are set in the DrawerNavGroup.
     const {
-        expandIcon = <MatIcon name={'expand-more'} size={24} color={theme.colors.text} />,
-        collapseIcon = <MatIcon name={'expand-less'} size={24} color={theme.colors.text} />,
+        // Inheritable Props
+        /* eslint-disable @typescript-eslint/no-unused-vars */
+        activeItemBackgroundColor,
+        activeItemBackgroundShape,
+        activeItemFontColor,
+        activeItemIconColor,
+        backgroundColor,
+        chevron,
+        collapseIcon,
+        disableActiveItemParentStyles,
+        divider,
+        expandIcon,
+        hidePadding,
+        itemFontColor,
+        itemIconColor,
+        nestedBackgroundColor,
+        nestedDivider,
+        theme: themeOverride,
+        /* eslint-enable @typescript-eslint/no-unused-vars */
+        // Drawer-specific props
+        activeItem,
+        onItemSelect,
+        styles = {},
+        // Other View Props
+        style,
+        ...viewProps
     } = props;
+
+    const theme = useTheme(themeOverride);
     const insets = useSafeAreaInsets();
     const defaultStyles = makeStyles(props, theme, insets);
 
-    const findChildByType = useCallback(
-        (type: string): JSX.Element[] =>
-            React.Children.map(props.children, (child: any) => {
-                if (child && child.type) {
-                    const name = child.type.displayName;
-                    if (name && name.includes(type)) {
-                        return child;
-                    }
-                }
-            }) || [],
-        [props]
-    );
-
     const getSectionByDisplayName = useCallback(
         (displayName: string, inherit = false): JSX.Element[] =>
-            findChildByType(displayName)
+            findChildByType(props.children, [displayName])
                 .slice(0, 1)
                 .map((child) => {
                     let inheritableProps = {};
                     if (inherit) {
-                        inheritableProps = inheritDrawerProps(
-                            {
-                                ...props,
-                                expandIcon,
-                                collapseIcon,
-                                // Set theme-related default props here.
-                                activeItemBackgroundColor:
-                                    props.activeItemBackgroundColor ||
-                                    color(theme.colors.primary).fade(0.95).rgb().string(),
-                                activeItemFontColor: props.activeItemFontColor || theme.colors.primary,
-                                activeItemIconColor: props.activeItemIconColor || theme.colors.primary,
-                                itemFontColor: props.itemFontColor || theme.colors.text,
-                                itemIconColor: props.itemIconColor || theme.colors.text,
-                            },
-                            child.props
-                        );
+                        inheritableProps = inheritSharedProps({ ...props, theme }, child.props);
                     }
                     return React.cloneElement(child, inheritableProps);
                 }),
@@ -87,12 +84,19 @@ export const Drawer: React.FC<DrawerProps> = (props) => {
     );
 
     return (
-        <Surface style={[defaultStyles.root, style]} {...viewProps}>
-            {getSectionByDisplayName('DrawerHeader')}
-            {getSectionByDisplayName('DrawerSubheader')}
-            {getSectionByDisplayName('DrawerBody', true)}
-            {getSectionByDisplayName('DrawerFooter')}
-        </Surface>
+        <DrawerContext.Provider
+            value={{
+                activeItem,
+                onItemSelect,
+            }}
+        >
+            <Surface style={[defaultStyles.root, styles.root, style]} {...viewProps}>
+                {getSectionByDisplayName('DrawerHeader')}
+                {getSectionByDisplayName('DrawerSubheader')}
+                {getSectionByDisplayName('DrawerBody', true)}
+                {getSectionByDisplayName('DrawerFooter')}
+            </Surface>
+        </DrawerContext.Provider>
     );
 };