+++ components/src/core/empty-state/empty-state.tsx
@@ -1,27 +1,37 @@
 import React, { ComponentType, useCallback } from 'react';
-import { View, StyleSheet, StyleProp, ViewStyle, TextStyle, ViewProps } from 'react-native';
+import { View, StyleSheet, StyleProp, ViewStyle, TextStyle, ViewProps, PixelRatio } from 'react-native';
 import { useTheme } from 'react-native-paper';
 import { H6, Subtitle2 } from '../typography';
 import { $DeepPartial } from '@callstack/react-theme-provider';
+import { WrapIconProps } from '../icon-wrapper';
 
-const defaultStyles = StyleSheet.create({
-    root: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        padding: 10,
-    },
-    title: {
-        textAlign: 'center',
-        marginTop: 16,
-    },
-    subtitle: {
-        textAlign: 'center',
-    },
-    actions: {
-        marginTop: 10,
-    },
-});
+type EmptyStateStyles = {
+    root?: ViewStyle;
+    title?: TextStyle;
+    description?: TextStyle;
+    actions?: ViewStyle;
+};
+const makeStyles = (theme: ReactNativePaper.Theme, fontScale: number): StyleSheet.NamedStyles<EmptyStateStyles> =>
+    StyleSheet.create({
+        root: {
+            flex: 1,
+            justifyContent: 'center',
+            alignItems: 'center',
+            padding: 16,
+        },
+        title: {
+            textAlign: 'center',
+            marginTop: 16 * fontScale,
+        },
+        description: {
+            // @ts-ignore
+            color: theme.dark ? theme.colors.textSecondary : theme.colors.text,
+            textAlign: 'center',
+        },
+        actions: {
+            marginTop: 16 * fontScale,
+        },
+    });
 
 export type EmptyStateProps = ViewProps & {
     /* Primary text to display */
@@ -31,7 +41,7 @@ export type EmptyStateProps = ViewProps & {
     description?: string;
 
     /* Icon to display */
-    IconClass?: ComponentType<{ size: number; color: string }>;
+    IconClass?: ComponentType<WrapIconProps>;
 
     /** Props to pass to the Icon component */
     IconProps?: { size?: number; color?: string };
@@ -79,6 +89,7 @@ export const EmptyState: React.FC<EmptyStateProps> = (props) => {
         ...viewProps
     } = props;
     const theme = useTheme(themeOverride);
+    const defaultStyles = makeStyles(theme, PixelRatio.getFontScale());
 
     const normalizeIconSize = useCallback((): number => {
         if (!iconSize) return 100;
@@ -87,7 +98,7 @@ export const EmptyState: React.FC<EmptyStateProps> = (props) => {
 
     const getColor = useCallback(
         (color: string | undefined): string => {
-            if (!color) return theme.colors.text;
+            if (!color) return theme.colors.textSecondary || theme.colors.text;
             if (Object.keys(theme.colors).indexOf(color) >= 0)
                 return theme.colors[color as keyof ReactNativePaper.Theme['colors']];
             return color;
@@ -106,9 +117,7 @@ export const EmptyState: React.FC<EmptyStateProps> = (props) => {
             {getIcon()}
             <H6 style={[defaultStyles.title, styles.title]}>{title}</H6>
             {description ? (
-                <Subtitle2 color={'primary'} style={[defaultStyles.subtitle, styles.description]}>
-                    {description}
-                </Subtitle2>
+                <Subtitle2 style={[defaultStyles.description, styles.description]}>{description}</Subtitle2>
             ) : null}
             {actions ? <View style={[defaultStyles.actions, styles.actions]}>{actions}</View> : null}
         </View>