+++ components/src/core/header/constants.ts
@@ -1,7 +1,10 @@
+import { PixelRatio } from 'react-native';
 import { getStatusBarHeight } from 'react-native-status-bar-height';
 
-export const EXTENDED_HEIGHT = 200 + getStatusBarHeight(true);
-export const REGULAR_HEIGHT = 56 + getStatusBarHeight(true);
+const fontScale = PixelRatio.getFontScale();
+
+export const EXTENDED_HEIGHT = 200 * fontScale + getStatusBarHeight(true);
+export const REGULAR_HEIGHT = 56 * fontScale + getStatusBarHeight(true);
 export const ICON_SIZE = 24;
 export const ICON_SPACING = 16;
 export const ANIMATION_LENGTH = 300;