+++ components/src/core/header/headerContent.tsx
@@ -9,6 +9,7 @@ import {
     TextStyle,
     I18nManager,
     Platform,
+    PixelRatio,
 } from 'react-native';
 import color from 'color';
 import { EXTENDED_HEIGHT, REGULAR_HEIGHT, ICON_SIZE, ICON_SPACING } from './constants';
@@ -21,6 +22,13 @@ const headerContentStyles = StyleSheet.create({
         flex: 1,
         flexDirection: 'column',
         justifyContent: 'flex-end',
+    },
+    searchContainer: {
+        flex: 1,
+        flexDirection: 'column',
+        justifyContent: 'center',
+        marginRight: 56,
+        marginTop: 0,
     },
 });
 
@@ -197,7 +205,10 @@ export type HeaderContentProps = {
     /** Optional header third line of text (hidden when collapsed) */
     info?: string;
 
-    actionCount?: number;
+    actionCount?: {
+        avatars: number;
+        icons: number;
+    };
 
     styles?: {
         root?: StyleProp<ViewStyle>;
@@ -211,9 +222,10 @@ export type HeaderContentProps = {
 };
 
 export const HeaderContent: React.FC<HeaderContentProps> = (props) => {
-    const { title, subtitle, info, actionCount = 0, theme, styles = {} } = props;
+    const { title, subtitle, info, actionCount = { avatars: 0, icons: 0 }, theme, styles = {} } = props;
     const { headerHeight } = useHeaderHeight();
     const { searching, searchConfig } = useSearch();
+    const fontScale = PixelRatio.getFontScale();
     const defaultStyles = headerContentStyles;
 
     let content: JSX.Element[] = [];
@@ -222,33 +234,35 @@ export const HeaderContent: React.FC<HeaderContentProps> = (props) => {
         content = [<SearchContent key={'search-content'} theme={theme} style={styles.search} />];
     } else {
         content = [
-            <HeaderTitle title={title} key="title_key" theme={theme} style={styles.title} />,
-            <HeaderInfo info={info} key="info_key" theme={theme} style={styles.info} />,
-            <HeaderSubtitle subtitle={subtitle} key="subtitle_key" theme={theme} style={styles.subtitle} />,
+            <HeaderTitle title={title} key="title_key" theme={theme} style={[styles.title]} />,
+            <HeaderInfo info={info} key="info_key" theme={theme} style={[styles.info]} />,
+            <HeaderSubtitle subtitle={subtitle} key="subtitle_key" theme={theme} style={[styles.subtitle]} />,
         ];
     }
 
     const getActionPanelWidth = useCallback(() => {
-        let length = actionCount;
-        if (searchConfig) length++;
-        if (length < 1) return 0;
-        length = Math.min(3, length);
-        return length * (ICON_SIZE + ICON_SPACING);
-    }, [actionCount, searchConfig]);
+        let iconLength = actionCount.icons;
+        const avatarLength = actionCount.avatars;
+
+        if (searchConfig) iconLength++;
+        if (iconLength + avatarLength < 1) return 0;
+        iconLength = Math.min(3 - avatarLength, iconLength);
+        return iconLength * (ICON_SIZE * fontScale + ICON_SPACING) + avatarLength * (40 * fontScale);
+    }, [actionCount, searchConfig, fontScale]);
 
     return (
         <Animated.View
             style={[
-                defaultStyles.titleContainer,
-                {
-                    marginRight:
-                        searchConfig && searching
-                            ? 56
-                            : headerHeight.interpolate({
-                                  inputRange: [REGULAR_HEIGHT, EXTENDED_HEIGHT],
-                                  outputRange: [getActionPanelWidth(), 0],
-                              }),
-                },
+                searching ? defaultStyles.searchContainer : defaultStyles.titleContainer,
+                searching
+                    ? {}
+                    : {
+                          marginRight: headerHeight.interpolate({
+                              inputRange: [REGULAR_HEIGHT, EXTENDED_HEIGHT],
+                              outputRange: [getActionPanelWidth(), 0],
+                          }),
+                          marginTop: subtitle && title ? 10 * fontScale : 0,
+                      },
                 styles.root,
             ]}
         >