+++ components/src/core/user-menu/user-menu.tsx
@@ -1,8 +1,9 @@
 import React, { useCallback, useState } from 'react';
-import { StyleSheet, TouchableWithoutFeedback, View, ViewStyle } from 'react-native';
+import { PixelRatio, StyleSheet, TouchableWithoutFeedback, View, ViewStyle } from 'react-native';
 import { BottomSheet } from './bottom-sheet';
 import { useTheme, Divider } from 'react-native-paper';
 import { InfoListItem, InfoListItemProps } from '../info-list-item/info-list-item';
+import * as Colors from '@pxblue/colors';
 
 export type UserMenuProps = {
     // Custom avatar to render as bottomsheet trigger
@@ -24,7 +25,9 @@ export type UserMenuProps = {
 };
 
 const useStyles = (
-    theme: ReactNativePaper.Theme
+    theme: ReactNativePaper.Theme,
+    fontScale: number,
+    avatarSize: number
 ): StyleSheet.NamedStyles<{
     root: ViewStyle;
     avatar: ViewStyle;
@@ -35,18 +38,29 @@ const useStyles = (
             backgroundColor: theme.colors.surface,
         },
         avatar: {
-            width: 40,
-            height: 40,
-            borderRadius: 40,
+            width: avatarSize * fontScale,
+            height: avatarSize * fontScale,
+            borderRadius: avatarSize * fontScale,
         },
         bottomsheet: {},
     });
 
 export const UserMenu: React.FC<UserMenuProps> = (props) => {
     const theme = useTheme();
-    const { avatar, backgroundColor, fontColor, iconColor, menuTitle, menuSubtitle, menuItems, styles = {} } = props;
+    const {
+        avatar,
+        backgroundColor,
+        fontColor,
+        iconColor = Colors.gray[500],
+        menuTitle,
+        menuSubtitle,
+        menuItems,
+        styles = {},
+    } = props;
+    const avatarSize = avatar.props.size || 40;
     const [showBottomSheet, setShowBottomSheet] = useState(false);
-    const defaultStyles = useStyles(theme);
+    const fontScale = PixelRatio.getFontScale();
+    const defaultStyles = useStyles(theme, fontScale, avatarSize);
 
     const openMenu = (): void => {
         if (menuItems) setShowBottomSheet(true);
@@ -55,6 +69,14 @@ export const UserMenu: React.FC<UserMenuProps> = (props) => {
     const closeMenu = (): void => {
         setShowBottomSheet(false);
     };
+
+    const getAvatar = useCallback(
+        () =>
+            React.cloneElement(avatar, {
+                size: avatarSize * fontScale,
+            }),
+        [avatar]
+    );
 
     const getMenu = useCallback(
         (): JSX.Element => (
@@ -66,7 +88,9 @@ export const UserMenu: React.FC<UserMenuProps> = (props) => {
                             title={menuTitle || ''}
                             subtitle={menuSubtitle}
                             leftComponent={
-                                <View style={[defaultStyles.avatar, styles.avatar, { marginLeft: 16 }]}>{avatar}</View>
+                                <View style={[defaultStyles.avatar, styles.avatar, { marginLeft: 16 }]}>
+                                    {getAvatar()}
+                                </View>
                             }
                             fontColor={fontColor}
                             backgroundColor={backgroundColor}
@@ -75,19 +99,33 @@ export const UserMenu: React.FC<UserMenuProps> = (props) => {
                     </>
                 )}
                 {menuItems &&
-                    menuItems.map((menuItem: InfoListItemProps, index: number) => (
-                        <InfoListItem
-                            {...menuItem}
-                            key={index}
-                            onPress={(): void => {
-                                closeMenu();
-                                if (menuItem.onPress) menuItem.onPress();
-                            }}
-                            iconColor={iconColor || menuItem.iconColor}
-                            fontColor={fontColor || menuItem.fontColor}
-                            backgroundColor={backgroundColor || menuItem.backgroundColor}
-                        />
-                    ))}
+                    menuItems.map((menuItem: InfoListItemProps, index: number) => {
+                        const menuItemStyles = menuItem.styles || {};
+                        return (
+                            <InfoListItem
+                                {...menuItem}
+                                key={index}
+                                onPress={(): void => {
+                                    closeMenu();
+                                    if (menuItem.onPress) menuItem.onPress();
+                                }}
+                                iconColor={iconColor || menuItem.iconColor}
+                                fontColor={fontColor || menuItem.fontColor}
+                                backgroundColor={backgroundColor || menuItem.backgroundColor}
+                                dense={menuItem.dense !== undefined ? menuItem.dense : true}
+                                styles={Object.assign(menuItemStyles, {
+                                    title: Object.assign(
+                                        {
+                                            fontSize: 16,
+                                            fontFamily: theme.fonts.regular.fontFamily,
+                                            fontWeight: theme.fonts.regular.fontWeight,
+                                        },
+                                        menuItemStyles.title
+                                    ),
+                                })}
+                            />
+                        );
+                    })}
             </>
         ),
         [menuItems, menuTitle, menuSubtitle, iconColor, fontColor, backgroundColor]
@@ -100,7 +138,7 @@ export const UserMenu: React.FC<UserMenuProps> = (props) => {
                 testID={'avatar'}
                 style={[defaultStyles.root, styles.root]}
             >
-                {avatar}
+                {getAvatar()}
             </TouchableWithoutFeedback>
             <BottomSheet
                 show={showBottomSheet}